requirements.txt
flask==3.0.1
openai==1.8.0
pyjwt==2.8.0
pip==23.3.2
requests==2.31.0
pnu-fortune==1.0.3
python-unsplash==1.2.5
